backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
python-dotenv>=1.0.0
pydantic>=2.8.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# AI Providers (install the one you need)
google-generativeai>=0.8.0  # For Gemini
anthropic>=0.18.0           # For Claude
openai>=1.12.0              # For OpenAI

# Testing
pytest>=8.0.0
pytest-asyncio>=0.21.1
httpx>=0.27.0
